+++ javaApp/SmartHome/src/main/java/smarthome/model/hardware/Termometr.java
@@ -41,7 +41,7 @@ public class Termometr extends Sensor{
      * @param max - max temperatura 
      * @param min - min temperatura 
      */
-    public Termometr(int ID, int room, int idPlytki, byte[] addres, Float temperatura, Float max, Float min) {
+    public Termometr(int ID, int room, int idPlytki, int[] addres, Float temperatura, Float max, Float min) {
         super(ID ,room, idPlytki, addres, SensorsTypes.THERMOMETR);
         this.temperatura = temperatura;
         this.max = max;
